requirements.txt
Djang==3.1
Pillow==7.2.0
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
django-appconf==1.0.4
django-autoslug==1.9.8
django-extensions==3.0.8
idna==2.10
pip==20.2.2
pytils==0.3
pytz==2020.1
requests==2.24.0
setuptools==49.6.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10